Existing directory-with-prefix behavior unit test
Introduces a new test to the testing suite that combines the existing
prefix option with a partial reference that uses a subdirectory. The aim
of this test of to ensure that the (arguably weird) behavior of applying
a prefix to the directory works in future versions of the plug-in after
changes are made to the prefix handling.